Update Updated the AddStep method for the ThermocyclingProtocolPlotManager service to be general for all ThermocyclingProtocolStepTypes and take ThermocyclingProtocolStep step instance as input. Some remaining bugs are the crashing when trying to remove a step from the protocol while highlighting the plot, loading a protocol does not refresh the plot (i.e. the plot does not clear prperly before the load). ThermocyclingProtocols can now be loaded and saved as XML files to and from the Thermocycling tab.
TemperatureUnits and TimeUnits with their defaults were added to Reseources

A FileSearchers class was added to utilities for loading and saving the protocol files.

As a note, the default directory for the loading and saving of protocols needs to be changed depending on the computer being used right now and should be updated to be in a config file.